# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kvarenda rental platform API for Uzbekistan
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Search and filter listings
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: rooms
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: amenities
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: has3dTour
          in: query
          schema:
            type: boolean
        - name: hasInsurance
          in: query
          schema:
            type: boolean
        - name: plan
          in: query
          schema:
            type: string
            enum: [basic, pro]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}/publish:
    post:
      operationId: publishListing
      tags: [listings]
      summary: Publish or unpublish a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                published:
                  type: boolean
              required: [published]
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: Get rental applications
      parameters:
        - name: listingId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit a rental application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/status:
    put:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Approve or reject an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                note:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /contracts:
    get:
      operationId: getContracts
      tags: [contracts]
      summary: Get contracts
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending_signatures, active, terminated]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

    post:
      operationId: createContract
      tags: [contracts]
      summary: Generate a contract from an approved application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/sign:
    post:
      operationId: signContract
      tags: [contracts]
      summary: Sign a contract (tenant or owner)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [tenant, owner]
              required: [role]
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerVerified:
          type: boolean
        title:
          type: string
        description:
          type: string
        address:
          type: string
        district:
          type: string
        rooms:
          type: integer
        floor:
          type: integer
        totalFloors:
          type: integer
        area:
          type: number
        priceUzs:
          type: integer
        deposit:
          type: integer
        plan:
          type: string
          enum: [basic, pro]
        status:
          type: string
          enum: [draft, pending_moderation, active, inactive]
        published:
          type: boolean
        verified:
          type: boolean
        has3dTour:
          type: boolean
        hasInsurance:
          type: boolean
        insuranceStatus:
          type: string
          enum: [none, available, insured]
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        rules:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        createdAt:
          type: string
      required: [id, ownerId, title, address, district, rooms, priceUzs, plan, status, published, verified, has3dTour, hasInsurance, insuranceStatus, createdAt]
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [listings, total, offset, limit]
    CreateListingBody:
      type: object
      properties:
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: string
        address:
          type: string
        district:
          type: string
        rooms:
          type: integer
        floor:
          type: integer
        totalFloors:
          type: integer
        area:
          type: number
        priceUzs:
          type: integer
        deposit:
          type: integer
        plan:
          type: string
          enum: [basic, pro]
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        rules:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required: [ownerId, title, address, district, rooms, priceUzs, plan]
    UpdateListingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        address:
          type: string
        district:
          type: string
        rooms:
          type: integer
        floor:
          type: integer
        totalFloors:
          type: integer
        area:
          type: number
        priceUzs:
          type: integer
        deposit:
          type: integer
        plan:
          type: string
          enum: [basic, pro]
        status:
          type: string
          enum: [draft, pending_moderation, active, inactive]
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        rules:
          type: string
        has3dTour:
          type: boolean
        hasInsurance:
          type: boolean
        insuranceStatus:
          type: string
          enum: [none, available, insured]
    Application:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        tenantVerified:
          type: boolean
        listingTitle:
          type: string
        district:
          type: string
        priceUzs:
          type: integer
        moveInDate:
          type: string
        durationMonths:
          type: integer
        purpose:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        note:
          type: string
        createdAt:
          type: string
      required: [id, listingId, tenantId, moveInDate, durationMonths, status, createdAt]
    CreateApplicationBody:
      type: object
      properties:
        listingId:
          type: integer
        tenantId:
          type: integer
        moveInDate:
          type: string
        durationMonths:
          type: integer
        purpose:
          type: string
        message:
          type: string
      required: [listingId, tenantId, moveInDate, durationMonths]
    Contract:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        listingId:
          type: integer
        tenantId:
          type: integer
        ownerId:
          type: integer
        tenantName:
          type: string
        ownerName:
          type: string
        listingTitle:
          type: string
        address:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        monthlyRentUzs:
          type: integer
        depositUzs:
          type: integer
        serviceFeePercent:
          type: number
        status:
          type: string
          enum: [draft, pending_signatures, active, terminated]
        tenantSigned:
          type: boolean
        ownerSigned:
          type: boolean
        tenantSignedAt:
          type: string
        ownerSignedAt:
          type: string
        terms:
          type: string
        createdAt:
          type: string
      required: [id, listingId, tenantId, ownerId, startDate, endDate, monthlyRentUzs, depositUzs, serviceFeePercent, status, tenantSigned, ownerSigned, createdAt]
    CreateContractBody:
      type: object
      properties:
        applicationId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        terms:
          type: string
      required: [applicationId, startDate, endDate]
...