# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kvarenda rental platform API for Uzbekistan
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  listingId: number;
  tenantId: number;
  moveInDate: string;
  durationMonths: number;
  purpose?: string;
  message?: string;
}