# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kvarenda rental platform API for Uzbekistan
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractBody {
  applicationId: number;
  startDate: string;
  endDate: string;
  terms?: string;
}